norm almost done

+++ src/redirection/redirection.c
@@ -6,13 +6,13 @@
 /*   By: ommohame < ommohame@student.42abudhabi.ae> +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/11 02:15:04 by ommohame          #+#    #+#             */
-/*   Updated: 2022/08/20 12:44:07 by ommohame         ###   ########.fr       */
+/*   Updated: 2022/08/24 00:14:03 by ommohame         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/minishell.h"
 
-int	redir_out(t_redir redir, int f)
+static int	redir_out(t_redir redir, int f)
 {
 	int		fd;
 
@@ -29,71 +29,7 @@ int	redir_out(t_redir redir, int f)
 	}
 }
 
-int		redir_in1(t_redir redir, int f)
-{
-	int		fd;
-
-	if (access(redir.file, F_OK) == -1)
-	{
-		ft_printf("minishell: %s: No such file or directory\n", redir.file);
-		return (-1);
-	}
-	if (access(redir.file, R_OK) == -1)
-	{
-		ft_printf("minishell: %s: Permission denied\n", redir.file);
-		return (-1);
-	}
-	fd = open(redir.file, O_RDONLY);
-	if (f == 1)
-		return (fd);
-	close (fd);
-	return (-1);
-}
-
-int		heredoc(t_redir redir, int f)
-{
-	int		fd;
-	char 	*tmp1;
-
-	fd = -1;
-	if (f == 1)
-		fd = open("./src/redirection/.heredoc.txt", O_WRONLY | O_TRUNC | O_CREAT, 0644);
-	while (1)
-	{
-		tmp1 = readline(">");
-		if (!tmp1)
-			return (-2);
-		if (!ft_strncmp(tmp1, redir.file, ft_strlen(tmp1)) && ft_strncmp(tmp1, "", 1)
-			&& (ft_strlen(tmp1) == ft_strlen(redir.file)))
-		{
-			free(tmp1);
-			close(fd);
-			if (f == 1)
-			{
-				fd = open("./src/redirection/.heredoc.txt", O_RDONLY);
-				return (fd);
-			}
-			else
-				return (-1);
-		}
-		ft_putstr_fd(tmp1, fd);
-		ft_putchar_fd('\n', fd);
-		free(tmp1);
-	}
-	if (f == 1)
-		return (fd);
-	return (-1);
-}
-
-int		redir_in(t_redir redir, int f)
-{
-	if (redir.type == 1)
-		return (redir_in1(redir, f));
-	else
-		return (heredoc(redir, f));
-}
-
-int		check_lastredir(t_redir redir)
+static int	check_lastredir(t_redir redir)
 {
 	int		fd;
 
@@ -113,7 +49,7 @@ int		check_lastredir(t_redir redir)
 	return (1);
 }
 
-int	redirect(t_line *line, int fd_in, int fd_out)
+static int	redirect(t_line *line, int fd_in, int fd_out)
 {
 	int	in;
 	int	out;
@@ -140,21 +76,17 @@ int	redirect(t_line *line, int fd_in, int fd_out)
 	return (1);
 }
 
-int	redirection(t_line *line)
-{
+static void	redir_loop(t_line *line)
+{	
 	int		f;
 	int		fd_in;
 	int		fd_out;
-	size_t	i;
 	t_redir	*redirr;
 
 	f = 0;
-	i = 0;
 	fd_in = -1;
 	fd_out = -1;
-	if (!line->cmd->redir || line->cmd->nredir == 0)
-		return (0);
-	while (i < line->cmd->nredir && line->cmd->redir)
+	while (line->cmd->redir)
 	{
 		if (check_lastredir(*line->cmd->redir))
 			f = 1;
@@ -167,9 +99,16 @@ int	redirection(t_line *line)
 		free(redirr->file);
 		free(redirr);
 		f = 0;
-		i++;
 	}
 	redirect(line, fd_in, fd_out);
-	unlink("./src/redirection/.heredoc.txt");
+}
+
+int	redirection(t_line *line)
+{
+	if (!line->cmd->redir || line->cmd->nredir == 0)
+		return (0);
+	redir_loop(line);
+	if (access("./src/redirection/.heredoc.txt", F_OK))
+		unlink("./src/redirection/.heredoc.txt");
 	return (1);
 }
